Prevent file overwrite by returning early if the file already exists

+++ src/mg/razherana/generator/file/FileGenerator.java
@@ -161,6 +161,8 @@ public class FileGenerator {
   }
 
   public void write() {
+    if(file.exists())
+      return;
     try (java.io.FileWriter fileWriter = new java.io.FileWriter(file)) {
       fileWriter.write(generate());
     } catch (Exception e) {
